requirements.txt
google-generativeai
google-auth-oauthlib
google-auth-httplib2
SQLAlchemy
python-telegram-bot
discord.py
pydantic
aiohttp
python-dotenv
schedule
psutil
beautifulsoup4
glob2
